fix: extract vnode to util
